+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * **BREAKING:** `parser.empty` is now a function (`parser.empty()`)
 * `Parsimmon.Parser.empty()` is a copy of `parser.empty()`
 * Adds `Parsimmon.isParser`
-* Documentation for `Parsimmon.formatError`
+* Adds `parser.fallback`
+* Documentation for `Parsimmon.formatError`, `Parsimmon.parse`
 
 ## version 0.9.2 (2016-08-07)